+++ linopy/__init__.py
@@ -7,5 +7,6 @@ Created on Wed Mar 10 11:03:06 2021
 """
 
 from linopy import model
+from linopy.expressions import merge
 from linopy.io import read_netcdf
 from linopy.model import LinearExpression, Model, Variable, available_solvers